+++ src/main/kotlin/g2901_3000/s2958_length_of_longest_subarray_with_at_most_k_frequency/Solution.kt
@@ -10,8 +10,8 @@ class Solution {
         var m1 = Int.MIN_VALUE
         var m2 = Int.MAX_VALUE
         for (num in nums) {
-            m1 = max(m1.toDouble(), num.toDouble()).toInt()
-            m2 = min(m2.toDouble(), num.toDouble()).toInt()
+            m1 = max(m1, num)
+            m2 = min(m2, num)
         }
         var max = 0
         val f = IntArray(m1 - m2 + 1)
