+++ src/main/kotlin/g3001_3100/s3074_apple_redistribution_into_boxes/Solution.kt
@@ -15,7 +15,7 @@ class Solution {
         var max = 0
         for (j in capacity) {
             count[j]++
-            max = max(max.toDouble(), j.toDouble()).toInt()
+            max = max(max, j)
         }
         for (i in max downTo 0) {
             if (count[i] >= 1) {
