+++ src/main/kotlin/g3001_3100/s3076_shortest_uncommon_substring_in_an_array/Solution.kt
@@ -28,8 +28,7 @@ class Solution {
                 val curLen = search(
                     cs,
                     i,
-                    min(m.toDouble(), (i + resultLen).toDouble())
-                        .toInt(),
+                    min(m, (i + resultLen)),
                     k,
                 )
                 if (curLen != -1) {
