+++ src/main/kotlin/g3201_3300/s3272_find_the_count_of_good_integers/Solution.kt
@@ -73,7 +73,7 @@ class Solution {
 
     private fun allKPalindromes(n: Int, k: Int): List<String> {
         val ans = StringBuilder(n)
-        ans.append("0".repeat(max(0.0, n.toDouble()).toInt()))
+        ans.append("0".repeat(max(0, n)))
         val rem = IntArray(n)
         rem[0] = 1
         for (i in 1 until n) {
