+++ src/main/kotlin/g3401_3500/s3459_length_of_longest_v_shaped_diagonal_segment/Solution.kt
@@ -31,7 +31,7 @@ class Solution {
             for (j in 0..<m) {
                 if (g[i][j] == 1) {
                     for (d in 0..3) {
-                        res = max(res.toDouble(), dp(i, j, 1, d, 1).toDouble()).toInt()
+                        res = max(res, dp(i, j, 1, d, 1))
                     }
                 }
             }
